refactor(knobs): relocate knob_space + knob_loader to src/knobs/
Move the knob search-space definition and CSV loader out of the PBT-specific
src/tuner/config/ package into the foundation-layer src/knobs/ package, where
they belong: neither module depends on anything in src/tuner, and both are
consumed by src/tuners/, src/evaluation/, and the BO baseline scripts.

- git mv knob_space.py, knob_loader.py -> src/knobs/ (history preserved)
- src/knobs/__init__.py becomes a façade re-exporting the public API
- src/tuner/config/__init__.py keeps re-exporting from src/knobs for the
  transitional PBT-internal import surface
- repoint all direct submodule importers (core, analysis, bo_baseline, tests)
- logger-color test expectations updated for the new module path

Part of ADR-006 tuners/ unification. Relocate-first; no behavior change.
