+++ src/knobs/__init__.py
@@ -1,5 +1,11 @@
 """PostgreSQL knob retrieval and analysis utilities."""
 
+from .knob_space import KnobSpace, KnobDefinition, KnobType, KnobScale
+from .knob_loader import (
+    load_knob_space_from_csv,
+    load_knob_space_for_tier,
+    get_knob_space,
+)
 from .retrieval import PostgreSQLKnobRetriever, KnobCategory, ConfigParameter
 from .policy import annotate_autotuning_policy, ensure_autotuning_policy_annotations
 from .preprocess_knobs import (
@@ -14,6 +20,13 @@ from .knob_metadata import (
 )
 
 __all__ = [
+    "KnobSpace",
+    "KnobDefinition",
+    "KnobType",
+    "KnobScale",
+    "load_knob_space_from_csv",
+    "load_knob_space_for_tier",
+    "get_knob_space",
     "PostgreSQLKnobRetriever",
     "KnobCategory",
     "ConfigParameter",
