+++ src/scripts/bo_baseline/runner.py
@@ -16,7 +16,7 @@ from smac.scenario import Scenario
 from smac.runhistory.dataclasses import TrialInfo, TrialValue
 from smac.runhistory.enumerations import StatusType
 
-from src.tuner.config import get_knob_space
+from src.knobs import get_knob_space
 from src.tuner.core.worker import Worker
 from src.tuner.benchmark.orchestrator import (
     WorkloadOrchestrator,
