+++ src/tuner/__init__.py
@@ -30,9 +30,9 @@ __author__ = "Data-Vanta"
 from src.tuner.core.worker import Worker
 
 from src.tuner.config.tuner_config import PBTConfig
-from src.tuner.config.knob_space import KnobSpace
+from src.knobs.knob_space import KnobSpace
 from src.tuner.core import GenerationResult
-from src.tuner.config.knob_loader import get_knob_space
+from src.knobs.knob_loader import get_knob_space
 
 __all__ = [
     "Worker",
